Move pks to a seperate accounts.go
